fix: standardize start over behavior, improve calltaker support
