+++ lib/components/narrative/trip-tools.js
@@ -8,6 +8,7 @@ import { Flag } from '@styled-icons/fa-solid/Flag'
 import { FormattedMessage, injectIntl } from 'react-intl'
 import { Print } from '@styled-icons/fa-solid/Print'
 import { Undo } from '@styled-icons/fa-solid/Undo'
+import { withRouter } from 'react-router'
 import bowser from 'bowser'
 import copyToClipboard from 'copy-to-clipboard'
 import PropTypes from 'prop-types'
@@ -15,6 +16,7 @@ import React, { Component } from 'react'
 
 import * as uiActions from '../../actions/ui'
 import { IconWithText } from '../util/styledIcon'
+import startOver from '../util/start-over'
 // Copy URL Button
 
 class CopyUrlButton extends Component {
@@ -181,6 +183,7 @@ class TripTools extends Component {
   render() {
     const {
       buttonTypes,
+      location,
       popupTarget,
       reactRouterConfig,
       reportConfig,
@@ -201,11 +204,10 @@ class TripTools extends Component {
           buttonComponents.push(<ReportIssueButton {...reportConfig} />)
           break
         case 'START_OVER':
-          // Determine "home" URL
-          let startOverUrl = '/'
-          if (reactRouterConfig && reactRouterConfig.basename) {
-            startOverUrl += reactRouterConfig.basename
-          }
+          const startOverUrl = startOver(
+            reactRouterConfig?.basename,
+            location?.search
+          )
           buttonComponents.push(
             // FIXME: The Spanish string does not fit in button width.
             <LinkButton
@@ -245,6 +247,7 @@ class TripTools extends Component {
 
 TripTools.propTypes = {
   buttonTypes: PropTypes.arrayOf(PropTypes.string),
+  location: PropTypes.object,
   popupTarget: PropTypes.string,
   reactRouterConfig: PropTypes.object,
   reportConfig: PropTypes.object,
@@ -264,4 +267,6 @@ const mapDispatchToProps = {
   setPopupContent: uiActions.setPopupContent
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(TripTools)
+export default withRouter(
+  connect(mapStateToProps, mapDispatchToProps)(TripTools)
+)
